State Machine Internals
Added a bmp.heartBeat() function that can be called often to keep
pressure data current. If it is never called, pressure data will still
never be more than 50 ms out of date.

State machine design eliminates the wait of up to 31 ms to return from
a call to readPressure(), most of which was spent in delay. This change
allows the loop to run at around 20kHz instead of about 30Hz max with
delays().

Application interface remains the same.