Fix description comments of functions
The description comments for the return types of
readRawTemperature() and readRawPressure() are
wrong. There is no point to add this information
in the comments anyway, as it is easily visible in
the function declaration. In addition it is easy
to change the type and to forget to update the
comment too.